pythonProject3/requirements.txt
streamlit
pdfplumber
scikit-learn
pyspellchecker
textstat
nltk